Implement RobustCCW and PolygonBuilder
RobustCCW uses ExpensiveCCW under the hood when ccw == 0. If after running
ExpensiveCCW and the sign of the determinant is still 0, resort to symbolic
perturbations to resolve the sign. This is made possible with an ExactFloat
implementation.

In order to test RobustCCW, colinear points are used in point_test.go. More
testing is needed here but it's a good start.

PolygonBuilder currently supports DIRECTED_XOR and can generate a slice of
loops (only tested with a single loop). Working on adding more Loop tests
with C++ API test data.